+++ packages/components-react/projects/remix/app/root.tsx
@@ -1,16 +1,22 @@
+import { type JSX, useState } from 'react';
 import type { MetaFunction } from '@remix-run/node';
-import { LiveReload, Meta, Outlet, Scripts, ScrollRestoration } from '@remix-run/react';
-import { componentsReady, PorscheDesignSystemProvider } from '@porsche-design-system/components-react/ssr';
+import { LiveReload, Meta, Outlet, Scripts, ScrollRestoration, useLocation, useNavigate } from '@remix-run/react';
+import { componentsReady, PorscheDesignSystemProvider, type Theme } from '@porsche-design-system/components-react/ssr';
 import { HeadPartials } from '~/head-partials.server';
 import { BodyPartials } from '~/body-partials.server';
+import { routes } from '~/routes';
 
-export const meta: MetaFunction = () => ([
+export const meta: MetaFunction = () => [
   { charset: 'utf-8' },
   { title: 'Porsche Design System - Remix' },
   { viewport: 'width=device-width,initial-scale=1' },
-]);
+];
 
 export default function App(): JSX.Element {
+  const navigate = useNavigate();
+  const [theme, setTheme] = useState<Theme>('light');
+  const themes: Theme[] = ['light', 'dark', 'auto'];
+
   return (
     <html lang="en">
       <head>
@@ -18,7 +24,22 @@ export default function App(): JSX.Element {
         {HeadPartials && <HeadPartials />}
       </head>
       <body>
-        <PorscheDesignSystemProvider>
+        <select value={useLocation().pathname} onChange={(e) => navigate(e.target.value)}>
+          <option disabled value="">
+            Select a page
+          </option>
+          {routes.map((route, i) => (
+            <option key={i} value={route.path} children={route.name} />
+          ))}
+        </select>
+
+        <select value={theme} onChange={(e) => setTheme(e.target.value as Theme)}>
+          {themes.map((item) => (
+            <option key={item} value={item} children={item} />
+          ))}
+        </select>
+
+        <PorscheDesignSystemProvider theme={theme}>
           <div id="app">
             <Outlet />
           </div>