+++ packages/components/src/components/pagination/pagination-styles.ts
@@ -88,7 +88,7 @@ export const getComponentCss = (
         display: 'flex',
         justifyContent: 'center',
         alignItems: 'center',
-        transition: ['color', 'border-color', 'background-color'].map(getTransition).join(), // for smooth transition between states
+        transition: `${getTransition('background-color')}, ${getTransition('border-color')}, ${getTransition('color')}`, // for smooth transition between states
         position: 'relative',
         width: buttonSize,
         height: buttonSize,