+++ packages/components/src/styles/link-button-styles.ts
@@ -115,7 +115,7 @@ export const getLinkButtonStyles = (
       backgroundColor,
       color: textColor,
       ...textSmallStyle,
-      transition: ['background-color', 'border-color', 'color'].map(getTransition).join(),
+      transition: `${getTransition('background-color')}, ${getTransition('border-color')}, ${getTransition('color')}`,
       ...buildResponsiveStyles(hideLabel, (hideLabelValue: boolean) => ({
         padding: hideLabelValue ? '13px' : '13px 26px',
         gap: hideLabelValue ? 0 : spacingStaticSmall,